+++ src/ScottPlot4/ScottPlot.Tests/PlotTypes/Scatter.cs
@@ -50,5 +50,23 @@ namespace ScottPlotTests.PlotTypes
             Assert.Greater(limits2.XMax, 10);
             Assert.Greater(limits2.YMax, 20);
         }
+
+        [Test]
+        public void Test_Scatter_Smooth()
+        {
+            double[] xs = { 18.5, 20.6, 22.3, 24.5, 26.6, 15, 15 };
+            double[] ys = { 1.43, 1.48, 1.6, 1.59, 1.53, 1.52, 1.6 };
+            var plt = new ScottPlot.Plot(600, 400);
+
+            var sp1 = plt.AddScatter(xs, ys, label: "DrawLines()");
+            sp1.Smooth = false;
+
+            var sp2 = plt.AddScatter(xs, ys, label: "DrawCurve()");
+            sp2.Smooth = true;
+
+            plt.Legend();
+            plt.SetAxisLimits(0, 30, 1.42, 1.62);
+            TestTools.SaveFig(plt);
+        }
     }
 }